[UPDATE] Added functionality of logging in the user, implemented security, authorised login requests [UPDATE] Added test methods to test for jwt token after logging in [ADD] Added various helper classes as JwtUtil, JwtRequestFilter, AuthenticationRequest, AuthenticationResponse required for jwt authentication

+++ src/main/java/com/sachet/postBook/model/BaseEntity.java
@@ -9,20 +9,20 @@ import javax.persistence.MappedSuperclass;
 public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     public BaseEntity() {
     }
 
-    public BaseEntity(int id) {
+    public BaseEntity(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
